pyproject.toml
[project]
name = "airbnb-concierge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler<4",
    "gcsa>=2.6.0",
    "imap-tools>=1.11.1",
    "ollama>=0.6.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
